+++ webdevproj/src/App.js
@@ -1,182 +1,11 @@
 import React, { useState } from "react";
 import { useForm } from "react-hook-form";
 import "./App.css";
-import {BrowserRouter as Router,Link,Route,Routes,useParams,useNavigate,} from "react-router-dom";
+import {BrowserRouter as Router,Link,Route,Routes,useParams, useNavigate,} from "react-router-dom";
+import LoginForm from "./LoginForm";  
+import SignupForm from "./SignupForm";  
+import Forum from "./Forum";  
 
-const LoginForm = ({ onLogin, onToggleSignup }) => {
-  const {
-    register,
-    handleSubmit,
-    formState: { errors },
-  } = useForm();
-
-  const onSubmit = (data) => {
-    onLogin(data.username);
-  };
-
-  return (
-    <div className="container">
-      <h2>Login</h2>
-      <form onSubmit={handleSubmit(onSubmit)}>
-        <label>Email:</label>
-        <input
-          type="text"
-          {...register("username", { required: "Username is required" })}
-        />
-        <br />
-        {errors.username && <p>{errors.username.message}</p>}
-        <br />
-        <label>Password:</label>
-        <input
-          type="password"
-          {...register("password", { required: "Password is required" })}
-        />
-        <br />
-        {errors.password && <p>{errors.password.message}</p>}
-        <br />
-        <button type="submit">Login</button>
-      </form>
-      <br />
-      <p>
-        Don't have an account? <button onClick={onToggleSignup}>Sign Up</button>
-      </p>
-    </div>
-  );
-};
-
-const SignupForm = ({ onSignup, onToggleSignup }) => {
-  const {
-    register,
-    handleSubmit,
-    formState: { errors },
-  } = useForm();
-
-  const onSubmit = (data) => {
-    onSignup(data.username);
-    console.log("Username:", data.username);
-    console.log("Password:", data.password);
-  };
-
-  return (
-    <div className="overlay">
-      <div className="container">
-        <h2>Sign Up</h2>
-        <form onSubmit={handleSubmit(onSubmit)}>
-          {" "}
-          {}
-          <label>Username:</label>
-          <input
-            type="text"
-            {...register("username", { required: "Username is required" })}
-          />
-          <br />
-          {errors.username && <p>{errors.username.message}</p>}
-          <br />
-          <label>Password:</label>
-          <input
-            type="password"
-            {...register("password", { required: "Password is required" })}
-          />
-          <br />
-          {errors.password && <p>{errors.password.message}</p>}
-          <br />
-          <button type="submit">Sign Up</button>
-        </form>
-        <br />
-        <p>
-          Already have an account? <span onClick={onToggleSignup}>Log In</span>
-        </p>
-      </div>
-    </div>
-  );
-};
-
-const ForumHome = ({ user }) => (
-  <div>
-    <h2>Welcome to the Community! {user}!</h2>
-    <p>Welcome to the TUP Community .</p>
-  </div>
-);
-
-const ForumProducts = () => {
-  const [products, setProducts] = useState([
-    {
-      id: 1,
-      name: "Product 1",
-      price: 19.99,
-      image: "https://www.pinterest.ph/pin/105130972538086806/",
-    },
-    {
-      id: 2,
-      name: "Product 2",
-      price: 29.99,
-      image: "https://example.com/product2.jpg",
-    },
-    {
-      id: 3,
-      name: "Product 3",
-      price: 39.99,
-      image: "https://example.com/product3.jpg",
-    },
-  ]);
-
-  return (
-    <div>
-      <h2>Products</h2>
-      <p>Explore the latest and discussions.</p>
-
-      <ul>
-        {products.map((product) => (
-          <li key={product.id}>
-            <img
-              src={product.image}
-              alt={product.name}
-              style={{ width: '50px', height: '50px', marginRight: '10px' }}
-            />
-            <strong>{product.name}</strong> - ${product.price}
-          </li>
-        ))}
-      </ul>
-    </div>
-  );
-};
-
-
-const ForumOrders = () => (
-  <div>
-    <h2>Orders</h2>
-    <p>View and your orders.</p>
-  </div>
-);
-
-const Forum = ({ user }) => {
-  const navigate = useNavigate();
-  const { id } = useParams();
-
-  return (
-    <div>
-      <nav>
-        <ul>
-          <li>
-            <Link to="/">Home</Link>
-          </li>
-          <li>
-            <Link to="/products">Products</Link>
-          </li>
-          <li>
-            <Link to="/orders">Orders</Link>
-          </li>
-        </ul>
-      </nav>
-
-      <Routes>
-        <Route path="/" element={<ForumHome user={user} />} />
-        <Route path="/products" element={<ForumProducts />} />
-        <Route path="/orders" element={<ForumOrders />} />
-      </Routes>
-    </div>
-  );
-};
 
 const App = () => {
   const [user, setUser] = useState(null);
