+++ webdevproj/src/ForumOrders.js
@@ -1,0 +1,10 @@
+import React, { useState } from "react";
+
+const ForumOrders = () => (
+    <div>
+      <h2>Orders</h2>
+      <p>View and your orders.</p>
+    </div>
+  );
+
+export default ForumOrders;
