+++ webdevproj/src/Forumhome.js
@@ -1,0 +1,10 @@
+import React from "react";
+
+const ForumHome = ({ user }) => (
+    <div>
+      <h2>Welcome to the Community! {user}!</h2>
+      <p>Welcome to the TUP Community .</p>
+    </div>
+  );
+  
+  export default ForumHome;
